Implement sidebar section
For the layout in this section, I relied heavily on CSS Grid to apply some of the more recent concepts covered in the past lessons.
The layout is a grid composed of three grid elements, the first being the dashboard, which is larger in size.
The second is the main navigation menu, which is a grid itself.
The third is the options menu, which is also a grid.